Maher v2.0
Panel de control casi completamente funcional, seguridad, y ajax.

+++ src/prueba/pruebaBundle/Controller/SolicitudController.php
@@ -25,13 +25,26 @@ class SolicitudController extends Controller
     public function indexAction(Request $request)
     {
         $session = $request->getSession();
-        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado' or $session->get('tipousuario') == 'Cliente') {
             $em = $this->getDoctrine()->getManager();
-
-            $solicituds = $em->getRepository('pruebaBundle:Solicitud')->findAll();
-
-            return $this->render('solicitud/index.html.twig', array(
+            /*
+             * con este condicional el cliente solo podra observar solo sus solicitudes nada mas
+             * permisologia evita que el cliente tenga acceso a las opciones del navbar
+             */
+            if ($session->get('tipousuario') == 'Cliente') {
+                $solicituds = $em->getRepository('pruebaBundle:Solicitud')->findBy(array('rif' => $session->get('rif')));
+                $permisologia = 0;
+            } else {
+                $solicituds = $em->getRepository('pruebaBundle:Solicitud')->findAll();
+                $permisologia = 1;
+            }
+            return $this->render('solicitud/index_sol.html.twig', array(
                         'solicituds' => $solicituds,
+                        'permisologia' => $permisologia,
+                        'nu'=>$session->get('nombreusuario'),
+                        'l'=>$session->get('login'),
+                        'diaactivo'=>$session->get('diaactivo'),
+                        
             ));
         } else {
             $this->get('session')->getFlashBag()->add(
@@ -50,22 +63,40 @@ class SolicitudController extends Controller
     public function newAction(Request $request)
     {
         $session = $request->getSession();
-        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado' or $session->get('tipousuario') == 'Cliente') {
+            
             $solicitud = new Solicitud();
             $form = $this->createForm('prueba\pruebaBundle\Form\SolicitudType', $solicitud);
             $form->handleRequest($request);
 
             if ($form->isSubmitted() && $form->isValid()) {
                 $em = $this->getDoctrine()->getManager();
+                /*
+                 * aqui se empieza a colocar los datos a traves del controlador para evitar que el usuario
+                 * ingrese datos errados
+                 */  
+               $solicitud->setEstatus('Activa'); 
+                switch ($solicitud->getTipopago()) {  //aqui asigna la prioridad de la solicitud por el tipo de pago
+                    case 'Efectivo':$solicitud->setPrioridad('Alta');
+                        break;
+                    case 'Cheque':$solicitud->setPrioridad('Media');
+                        break;
+                    case 'transferencia':$solicitud->setPrioridad('Media');
+                        break;
+                    case 'Credito':$solicitud->setPrioridad('Baja');
+                        break;
+                }                
                 $em->persist($solicitud);
                 $em->flush();
 
                 return $this->redirectToRoute('solicitud_show', array('id' => $solicitud->getIdsolicitud()));
             }
 
-            return $this->render('solicitud/new.html.twig', array(
+            return $this->render('solicitud/new_sol.html.twig', array(
                         'solicitud' => $solicitud,
                         'form' => $form->createView(),
+                        'rif'=>$session->get('rif'),
+                       
             ));
         } else {
             $this->get('session')->getFlashBag()->add(
@@ -78,18 +109,48 @@ class SolicitudController extends Controller
     /**
      * Finds and displays a Solicitud entity.
      *
-     * @Route("/{id}", name="solicitud_show")
+     * @Route("/{idsolicitud}", name="solicitud_show")
      * @Method("GET")
      */
-    public function showAction(Solicitud $solicitud)
+    public function showAction(Solicitud $solicitud,Request $request,$idsolicitud)
     {
         $session = $request->getSession();
-        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
-            $deleteForm = $this->createDeleteForm($solicitud);
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado' or $session->get('tipousuario') == 'Cliente') {
+            /**
+             * con este find se busca los detalles ya ingresados en la solicitud actual
+             * para ser mostrados en el formulario de ingreso
+             */
+            $em = $this->getDoctrine()->getManager();
+            $findsd = $em->createQuery(
+                       'select sd.idsolicituddetalle,p.nombreproducto,sd.cantidad,sd.precio,sd.total
 
+                        from pruebaBundle:producto p inner join pruebaBundle:solicituddetalle sd with
+                        sd.idproducto=p.idproducto
+
+                        where sd.idsolicitud=' . $idsolicitud . ''
+            );
+            $resultadosd=$findsd->getResult();
+            //***********************************************************************
+            /**
+             * con este find se busca la solicitud actual
+             * para ser mostrados en el formulario de ingreso
+             */
+            
+            $finds = $em->createQuery(
+                       'select s.fechaentrega,s.estatus,s.tipopago,s.prioridad,sum(sd.total) t
+
+                        from pruebaBundle:Solicitud s inner join pruebaBundle:Solicituddetalle sd with
+                        sd.idsolicitud=s.idsolicitud                        
+
+                        where s.idsolicitud=' . $idsolicitud . ''
+            );
+            $resultados=$finds->getResult();
+            //***********************************************************************S
             return $this->render('solicitud/show.html.twig', array(
-                        'solicitud' => $solicitud,
-                        'delete_form' => $deleteForm->createView(),
+                        'idsolicitud' => $idsolicitud,
+                        'solicituddetalles'=>$resultadosd,
+                        'solicitud'=>$resultados,
+                        
             ));
         } else {
             $this->get('session')->getFlashBag()->add(
@@ -108,7 +169,7 @@ class SolicitudController extends Controller
     public function editAction(Request $request, Solicitud $solicitud)
     {
         $session = $request->getSession();
-        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado' or $session->get('tipousuario') == 'Cliente') {
             $deleteForm = $this->createDeleteForm($solicitud);
             $editForm = $this->createForm('prueba\pruebaBundle\Form\SolicitudType', $solicitud);
             $editForm->handleRequest($request);
@@ -121,7 +182,7 @@ class SolicitudController extends Controller
                 return $this->redirectToRoute('solicitud_edit', array('id' => $solicitud->getIdsolicitud()));
             }
 
-            return $this->render('solicitud/edit.html.twig', array(
+            return $this->render('solicitud/edit_sol.html.twig', array(
                         'solicitud' => $solicitud,
                         'edit_form' => $editForm->createView(),
                         'delete_form' => $deleteForm->createView(),
@@ -141,24 +202,43 @@ class SolicitudController extends Controller
      * @Method("DELETE")
      */
     public function deleteAction(Request $request, Solicitud $solicitud)
-    {
+ {
+        /*
+         * con el siguiente condicional se especifica que un cliente aunque el haya
+         * hecho el registro de una solicitud el no puede eliminarla
+         */
         $session = $request->getSession();
-        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
-            $form = $this->createDeleteForm($solicitud);
-            $form->handleRequest($request);
-
-            if ($form->isSubmitted() && $form->isValid()) {
-                $em = $this->getDoctrine()->getManager();
-                $em->remove($solicitud);
-                $em->flush();
-            }
-
-            return $this->redirectToRoute('solicitud_index');
-        } else {
+        if ($session->get('tipousuario') == 'Cliente') {
             $this->get('session')->getFlashBag()->add(
-                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+                    'Alerta', "Usted no tiene permiso para eliminar una solicitud, llame a Distribuidora Maher, C.A. para que eliminen dicha solicitud"
             );
+            return $this->redirect($this->generateUrl('solicitud_index'));
+        } else {
+            if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+                $form = $this->createDeleteForm($solicitud);
+                $form->handleRequest($request);
+
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $em = $this->getDoctrine()->getManager();
+                    $em->remove($solicitud);
+                    $em->flush();
+                    $this->get('session')->getFlashBag()->add(
+                            'Mensaje', "El registro fue eliminado satisfactoriamente."
+                    );
+                } else {
+                    $this->get('session')->getFlashBag()->add(
+                            'Alerta', "El registro no pudo ser eliminado, puede que el registro este relacionado."
+                    );
+                }
+
+                return $this->redirectToRoute('solicitud_index');
+            } else {
+                $this->get('session')->getFlashBag()->add(
+                        'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+                );
+            }
         }
+
         return $this->redirect($this->generateUrl('inicio'));
     }
 
@@ -171,18 +251,31 @@ class SolicitudController extends Controller
      */
     private function createDeleteForm(Solicitud $solicitud)
     {
-        $session = $request->getSession();
-        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+        
             return $this->createFormBuilder()
                             ->setAction($this->generateUrl('solicitud_delete', array('id' => $solicitud->getIdsolicitud())))
                             ->setMethod('DELETE')
                             ->getForm()
             ;
-        } else {
-            $this->get('session')->getFlashBag()->add(
-                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
-            );
-        }
-        return $this->redirect($this->generateUrl('inicio'));
+        
+    }
+    /**
+     * en esta funcion de la clase se va a modificar el estatus de la solicitud para que el sistema lo tome en cuenta 
+     * cuando este logueado el administrador
+     * @Route("/{idsolicitud}/enviar", name="solicitud_enviar")
+     * @Method({"GET", "POST"})
+     */
+    public function enviarAction($idsolicitud){
+        $findsolicitud = $this->getDoctrine()
+                    ->getRepository('pruebaBundle:Solicitud')
+                    ->findOneBy(array('idsolicitud' => $idsolicitud));
+        $em = $this->getDoctrine()->getManager();
+        $findsolicitud->setEstatus('Enviada'); //con este estatus se envia la solicitud y se evita que el usuario la modifique
+        $em->persist($findsolicitud);
+        $em->flush();
+        $this->get('session')->getFlashBag()->add(
+                        'Mensaje', "Su solicitud fue confirmada y enviada a maher para su gestionamiento, gracias por preferirnos"
+                );
+        return $this->redirect($this->generateUrl('solicitud_index'));
     }
 }
